X-Ibition-2025-Robodog/INTERFACE 2025 EAC/requirements.txt
customtkinter>=5.2.0
opencv-python>=4.8.0
mediapipe>=0.10.0
ultralytics>=8.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
pyserial>=3.5
torch>=2.0.0
